backend/requirements.txt
Flask>=2.0
Flask-CORS>=3.0 # Add Flask-CORS
python-dotenv>=0.19
google-generativeai>=0.3
requests>=2.25 # Often useful, though not strictly needed for this exact code
celery[redis]>=5.0 # Adds celery and the redis dependency
redis>=4.0
waitress>=2.0 # Production WSGI server for Windows/Unix




celery -A tasks worker --pool=eventlet --loglevel=info
waitress-serve --host 127.0.0.1 --port 5001 app:app
npm run dev
